+++ files/lib/system/search/ConversationMessageSearch.class.php
@@ -1,6 +1,7 @@
 <?php
 namespace wcf\system\search;
 use wcf\data\conversation\message\SearchResultConversationMessageList;
+use wcf\data\conversation\Conversation;
 use wcf\form\IForm;
 use wcf\system\database\util\PreparedStatementConditionBuilder;
 use wcf\system\WCF;
@@ -17,10 +18,22 @@ use wcf\system\WCF;
  */
 class ConversationMessageSearch extends AbstractSearchableObjectType {
 	/**
+	 * id of the searched conversation
+	 * @var	integer
+	 */
+	public $conversationID = 0;
+	
+	/**
+	 * searched conversation
+	 * @var	Conversation
+	 */
+	public $conversation;
+	
+	/**
 	 * message data cache
 	 * @var	array<\wcf\data\conversation\message\SearchResultConversationMessage>
 	 */
-	public $messageCache = array();
+	public $messageCache = [];
 	
 	/**
 	 * @see	\wcf\system\search\ISearchableObjectType::cacheObjects()
@@ -32,6 +45,15 @@ class ConversationMessageSearch extends AbstractSearchableObjectType {
 		foreach ($messageList->getObjects() as $message) {
 			$this->messageCache[$message->messageID] = $message;
 		}
+	}
+	
+	/**
+	 * @see	\wcf\system\search\ISearchableObjectType::getAdditionalData()
+	 */
+	public function getAdditionalData() {
+		return [
+			'conversationID' => $this->conversationID
+		];
 	}
 	
 	/**
@@ -47,7 +69,7 @@ class ConversationMessageSearch extends AbstractSearchableObjectType {
 	 */
 	public function getJoins() {
 		return "JOIN wcf".WCF_N."_conversation_to_user conversation_to_user ON (conversation_to_user.participantID = ".WCF::getUser()->userID." AND conversation_to_user.conversationID = ".$this->getTableName().".conversationID)
-		LEFT JOIN wcf".WCF_N."_conversation conversation ON (conversation.conversationID = ".$this->getTableName().".conversationID)";
+			LEFT JOIN wcf".WCF_N."_conversation conversation ON (conversation.conversationID = ".$this->getTableName().".conversationID)";
 	}
 	
 	/**
@@ -78,6 +100,12 @@ class ConversationMessageSearch extends AbstractSearchableObjectType {
 		$conditionBuilder = new PreparedStatementConditionBuilder();
 		$conditionBuilder->add('conversation_to_user.hideConversation IN (0,1)');
 		
+		if (isset($_POST['conversationID'])) {
+			$this->conversationID = intval($_POST['conversationID']);
+			
+			$conditionBuilder->add('conversation.conversationID = ?', [$this->conversationID]);
+		}
+		
 		return $conditionBuilder;
 	}
 	
@@ -86,5 +114,37 @@ class ConversationMessageSearch extends AbstractSearchableObjectType {
 	 */
 	public function isAccessible() {
 		return (WCF::getUser()->userID ? true : false);
+	}
+	
+	/**
+	 * @see	\wcf\system\search\ISearchableObjectType::getFormTemplateName()
+	 */
+	public function getFormTemplateName() {
+		if ($this->conversation) {
+			return 'searchConversationMessage';
+		}
+		
+		return null;
+	}
+	
+	/**
+	 * @see	\wcf\system\search\ISearchableObjectType::show()
+	 */
+	public function show(IForm $form = null) {
+		// get existing values
+		if ($form !== null && isset($form->searchData['additionalData']['com.woltlab.wcf.conversation.message']['conversationID'])) {
+			$this->conversationID = $form->searchData['additionalData']['com.woltlab.wcf.conversation.message']['conversationID'];
+			
+			if ($this->conversationID) {
+				$this->conversation = Conversation::getUserConversation($this->conversationID, WCF::getUser()->userID);
+				
+				if ($this->conversation === null || !$this->conversation->canRead()) {
+					$this->conversationID = 0;
+					$this->conversation = null;
+				}
+			}
+		}
+		
+		WCF::getTPL()->assign('searchedConversation', $this->conversation);
 	}
 }