+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 ## 2.2
 
 ### 2.2.0 Alpha 1 (XXXX-YY-ZZ)
+
+* Only search active conversation when using search bar in a conversation.